Minor fixes

+++ mvc/src/main/java/com/mayada1994/moviewatchlist_mvc/fragments/WatchlistFragment.kt
@@ -33,6 +33,8 @@ class WatchlistFragment : Fragment() {
 
     private val selectedMovies: ArrayList<Movie> = arrayListOf()
 
+    private var deleteDialog: AlertDialog? = null
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -98,20 +100,24 @@ class WatchlistFragment : Fragment() {
     }
 
     private fun showDeleteMoviesDialog() {
-        val dialogView = DialogEditWatchlistBinding.inflate(layoutInflater)
-        val alertDialog = AlertDialog.Builder(requireContext()).setView(dialogView.root).create()
-
-        with(dialogView) {
-            txtPrompt.text = getString(R.string.dialog_delete_movie_prompt_message)
-            btnOk.setOnClickListener {
-                alertDialog.dismiss()
-                deleteMovies()
+        if (deleteDialog == null) {
+            val dialogView = DialogEditWatchlistBinding.inflate(layoutInflater)
+            val alertDialog =
+                AlertDialog.Builder(requireContext()).setView(dialogView.root).create()
+            deleteDialog = alertDialog
+            with(dialogView) {
+                txtPrompt.text = getString(R.string.dialog_delete_movie_prompt_message)
+                btnOk.setOnClickListener {
+                    alertDialog.dismiss()
+                    deleteMovies()
+                }
+                btnCancel.setOnClickListener {
+                    alertDialog.dismiss()
+                }
             }
-            btnCancel.setOnClickListener {
-                alertDialog.dismiss()
-            }
+            alertDialog.setOnDismissListener { deleteDialog = null }
+            alertDialog.show()
         }
-        alertDialog.show()
     }
 
     private fun getWatchlist() {
